Cast update due to the use of G++ flag -Wconversion and -Wfatal-errors
git-svn-id: http://130.66.124.6/svn/dev_pilotage/simulator_sdk/trunk/data_source@534 5574b880-4a8e-4d59-90b7-38b80912c59c

+++ data_source/inc/TypeCoercion.hpp
@@ -47,7 +47,7 @@ DECLARE_ARITHMETIC_TYPE(size_t);
 template <class T>
 typename enable_if<is_arithmetic<T>::value,void>::type coerce(std::list<double>& ret, const T thing_to_convert)
 {
-    ret.push_back(thing_to_convert);
+    ret.push_back((double)thing_to_convert);
 }
 
 template <typename T> void coerce(std::list<double>& ret, const std::vector<T>& thing_to_convert)
